Module P0_lib__P0_2021

Monadic parser combinators.

Return at most 1 parse at each point (so, alternatives are ordered, and the first applicable is taken).

A rewrite of 2017 version.

val iter_k : (k:('a -> 'b) -> 'a -> 'b) -> 'a -> 'b

type state = {

init_input : string;

len : int;

i : int;

debug : bool;

max : int Stdlib.ref;

}

val update_max : state -> state

val string_to_state : debug:bool -> string -> state

module M : sig ... end

include M

type 'a m

val return : 'a -> 'a m

val (>>=) : 'a m -> ('a -> 'b m) -> 'b m

val run : 'a m -> state -> ('a * state) option

val inject : (state -> ('a * state) option) -> 'a m

val fail : unit -> 'a m

val parse : debug:bool -> 'a m -> string -> 'a option

val get_state : unit -> state m

val set_state : state -> unit m

val debug : msg:string -> 'a m -> 'a m

Modify a parser so that it prints a message and the position in the input that has been reached before it tries to parse

val get_substring : int -> string m

val skip : int -> unit m

val _a : string -> unit m

val a : string -> string m

This version does not allocate substrings

val eps : string m

module Re_ : sig ... end

Link to ocaml-re; we need to match against a string from a certain position. This is a thin layer over ocaml-re. NOTE: you can use everything in Re in addition to the following.

val exec : Re.re -> string m

val opt : 'a m -> 'a option m

Try to parse; return None if not possible

val alt : 'a m -> 'a m -> 'a m

Prefer first option, but if not possible try second

val alt_list : 'a m list -> 'a m

val end_of_input : unit m

val then_ : 'a m -> 'b m -> ('a * 'b) m

Parse one then another; return a pair

val seq : 'a m -> 'b m -> ('a * 'b) m

seq is just then_

val seq_list : 'a m list -> 'a list m

val rep : 'a m -> 'a list m

Repetition

val list_with_sep : sep:'a m -> 'b m -> [> `Empty | `Not_empty of 'b * ('a * 'b) list ] m

This variant returns the matching seps as well

val list : sep:'a m -> 'b m -> 'b list m

This variant does not return the seps

val upto_a : string -> string m

val any_of : string list -> string m

Parse one of the given literals

val rep_any_but : string list -> string m

Parse until reaching one of the given literals

val ws : string m

val ws_nnl : string m

module Test : functor () sig ... end