Module P0_lib__P0_2021.Re_

Link to ocaml-re; we need to match against a string from a certain position. This is a thin layer over ocaml-re. NOTE: you can use everything in Re in addition to the following.

type t = Re.t

type re = Re.re

val compile : t -> re

val exec_opt : pos:int -> Re.re -> string -> int option

Return end position of matching substring (which starts at pos)

val exec_opt_s : pos:int -> Re.re -> string -> string option

Return matching substring

val exec : Re.re -> string m

Parser, returns matching substring

val exec_g : Re.re -> Re.Group.t m

Parser, returns matching group; doesn't advance the position

val a : string -> string m